modify data.m error

+++ cut/data.m
@@ -7,7 +7,7 @@ maskMatrix = CurrRes(:, :, 1);
 maskMatrix(maskMatrix ~= 0) = -1;
 
 user_mask = ~fixedBG;
-imwrite(uint8(user_mask),'images/user_mask.jpg');
+imwrite(user_mask,'images/user_mask.jpg');
 
 load imBounds.mat;
 imwrite(uint8(imBounds),'images/user_draw.jpg');
